+++ huanshankeji-team/gradle-library/src/main/kotlin/com/huanshankeji/team/gitversioning/gitlabpackageregistry/Repository.kt
@@ -4,6 +4,7 @@ import com.huanshankeji.GradleCommonExperimentalApi
 import com.huanshankeji.artifacts.MavenRepositoryHandlerContext
 import com.huanshankeji.artifacts.leadingProjectNameModuleRegex
 import com.huanshankeji.gitlab.packageregistry.maven.GITLAB_COM_HOST
+import com.huanshankeji.gitversioning.ConventionVersionRegexes
 import com.huanshankeji.gitversioning.gitlabpackageregistry.gitlabPackageRegistryProjectLevelEndpointConventionMavenRepositories
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.provider.ProviderFactory
@@ -11,36 +12,48 @@ import java.net.URI
 
 context(providers: ProviderFactory, _: (path: Any) -> URI)
 fun RepositoryHandler.huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String, groupRegex: String, moduleRegex: String
+    projectId: String,
+    groupRegex: String,
+    moduleRegex: String,
+    versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     gitlabPackageRegistryProjectLevelEndpointConventionMavenRepositories(
-        GITLAB_COM_HOST, projectId, groupRegex, moduleRegex
+        GITLAB_COM_HOST, projectId, groupRegex, moduleRegex, versionRegexes
     )
 
 @GradleCommonExperimentalApi
 fun MavenRepositoryHandlerContext.huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String, groupRegex: String, moduleRegex: String
+    projectId: String,
+    groupRegex: String,
+    moduleRegex: String,
+    versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     context(providers, uri) {
         repositories.huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-            projectId, groupRegex, moduleRegex
+            projectId, groupRegex, moduleRegex, versionRegexes
         )
     }
 
 context(providers: ProviderFactory, _: (path: Any) -> URI)
 fun RepositoryHandler.defaultHuanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String, groupRegex: String, projectName: String
+    projectId: String,
+    groupRegex: String,
+    projectName: String,
+    versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     huanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-        projectId, groupRegex, leadingProjectNameModuleRegex(projectName)
+        projectId, groupRegex, leadingProjectNameModuleRegex(projectName), versionRegexes
     )
 
 @GradleCommonExperimentalApi
 fun MavenRepositoryHandlerContext.defaultHuanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-    projectId: String, groupRegex: String, projectName: String
+    projectId: String,
+    groupRegex: String,
+    projectName: String,
+    versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     context(providers, uri) {
         repositories.defaultHuanshankejiGitlabPackageRegistryProjectEndpointConventionMavenRepositories(
-            projectId, groupRegex, projectName
+            projectId, groupRegex, projectName, versionRegexes
         )
     }