+++ huanshankeji-team/gradle-library/src/main/kotlin/com/huanshankeji/team/gitversioning/opensourcemavenconvention/githubpackages/Repository.kt
@@ -3,6 +3,7 @@ package com.huanshankeji.team.gitversioning.opensourcemavenconvention.githubpack
 import com.huanshankeji.GradleCommonExperimentalApi
 import com.huanshankeji.artifacts.MavenRepositoryHandlerContext
 import com.huanshankeji.artifacts.leadingProjectNameModuleRegex
+import com.huanshankeji.gitversioning.ConventionVersionRegexes
 import com.huanshankeji.gitversioning.opensourceconvention.githubpackages.githubPackagesSingleProjectOpenSourceConventionMavenRepositories
 import com.huanshankeji.team.HUANSHANKEJI_GROUP_REGEX
 import com.huanshankeji.team.HUANSHANKEJI_IN_LOWERCASE
@@ -17,10 +18,11 @@ context(providers: ProviderFactory, _: (path: Any) -> URI)
 fun RepositoryHandler.huanshankejiGithubPackagesOpenSourceMavenConventionProjectRepositories(
     projectName: String,
     groupRegex: String = HUANSHANKEJI_GROUP_REGEX,
-    moduleRegex: String = leadingProjectNameModuleRegex(projectName)
+    moduleRegex: String = leadingProjectNameModuleRegex(projectName),
+    versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     githubPackagesSingleProjectOpenSourceConventionMavenRepositories(
-        HUANSHANKEJI_IN_LOWERCASE, projectName, groupRegex, moduleRegex
+        HUANSHANKEJI_IN_LOWERCASE, projectName, groupRegex, moduleRegex, versionRegexes
     )
 
 @GradleCommonExperimentalApi
@@ -28,9 +30,10 @@ fun MavenRepositoryHandlerContext.huanshankejiGithubPackagesOpenSourceMavenConve
     projectName: String,
     groupRegex: String = HUANSHANKEJI_GROUP_REGEX,
     moduleRegex: String = leadingProjectNameModuleRegex(projectName),
+    versionRegexes: ConventionVersionRegexes = ConventionVersionRegexes(),
 ) =
     context(providers, uri) {
         repositories.huanshankejiGithubPackagesOpenSourceMavenConventionProjectRepositories(
-            projectName, groupRegex, moduleRegex
+            projectName, groupRegex, moduleRegex, versionRegexes
         )
     }