+++ kotlin-common/gradle-library/src/main/kotlin/com/huanshankeji/gitversioning/VersionMatching.kt
@@ -11,10 +11,15 @@ fun isDevCommitVersion(version: String): Boolean =
 fun Project.isDevCommitVersion(): Boolean =
     isDevCommitVersion(version.toString())
 
-const val DEV_COMMIT_VERSION_REGEX = """.*-dev-commit-[0-9a-f]+"""
+const val DEV_COMMIT_VERSION_REGEX_SUFFIX = """-dev-commit-[0-9a-f]+"""
+
+const val DEV_COMMIT_VERSION_REGEX = """.*$DEV_COMMIT_VERSION_REGEX_SUFFIX"""
+
+fun devCommitVersionRegexOf(releaseVersionRegex: String) =
+    """(?:$releaseVersionRegex)$DEV_COMMIT_VERSION_REGEX_SUFFIX"""
 
 /** Clean committed dev build, e.g. `1.0.0-dev-commit-abc123`. */
-const val STANDARD_DEV_COMMIT_VERSION_REGEX = """$STANDARD_RELEASE_VERSION_REGEX-dev-commit-[0-9a-f]+"""
+val STANDARD_DEV_COMMIT_VERSION_REGEX = devCommitVersionRegexOf(STANDARD_RELEASE_VERSION_REGEX)
 
 
 fun isDirtyDevCommitVersion(version: String): Boolean =
@@ -25,9 +30,9 @@ fun Project.isDirtyDevCommitVersion(): Boolean =
     isDirtyDevCommitVersion(version.toString())
 
 /** Matches [SNAPSHOT_VERSION_REGEX] or [STANDARD_DEV_COMMIT_VERSION_REGEX]. */
-const val SNAPSHOT_AND_STANDARD_DEV_COMMIT_VERSION_REGEX = "$SNAPSHOT_VERSION_REGEX|$STANDARD_DEV_COMMIT_VERSION_REGEX"
+val SNAPSHOT_AND_STANDARD_DEV_COMMIT_VERSION_REGEX = "$SNAPSHOT_VERSION_REGEX|$STANDARD_DEV_COMMIT_VERSION_REGEX"
 
 
 /** Matches [STANDARD_DEV_COMMIT_VERSION_REGEX] or [STANDARD_RELEASE_VERSION_REGEX]. Not used now. */
-const val STANDARD_DEV_COMMIT_AND_RELEASE_VERSION_REGEX =
+val STANDARD_DEV_COMMIT_AND_RELEASE_VERSION_REGEX =
     "$STANDARD_DEV_COMMIT_VERSION_REGEX|$STANDARD_RELEASE_VERSION_REGEX"